[automation/nodejs] Add Workspace.outputs
